+++ FT64/trunk/software/C64/source/Intexpr.cpp
@@ -25,7 +25,7 @@
 //
 #include "stdafx.h"
 
-int GetIntegerExpression(ENODE **pnode)       /* simple integer value */
+int64_t GetIntegerExpression(ENODE **pnode)       /* simple integer value */
 { 
 	TYP *tp;
 	ENODE *node;
